refactor: use best go practices for project structure
